Harden quantities, credit lifecycle and refund allocations (H1-H3)
Static-review fixes for the three high-severity audit findings, pending
real execution (migration e8b2c47f91a3 is written but NOT applied).

H1 - Quantities are strictly positive integers end to end: order_lines,
order_line_modifiers and refund allocations move to INTEGER with
CHECK >= 1, API schemas use strict ints (no coercion of "1", 1.0 or
booleans), and every silent int() truncation in pricing is gone - a
fractional credits line can no longer produce a free product. The
migration refuses to run against fractional data instead of truncating.

H2 - Redemption lifecycle is now explicit and DB-enforced: reserved
redemptions can only be released by cancellation (never refunded),
consumed ones refund up to what was actually spent, earn reversals only
exist for completed orders, and four partial unique indexes plus a
refund_allocation_id causal link make every credit compensation
idempotent under retries. Credits require a customer: new CHECK zeroes
the possibility of earning or redeeming on orders without
customer_user_id, and internal channels must record the capturing
employee (source='online' OR created_by IS NOT NULL).

H3 - Per-line refund caps accumulate across refunds inside the
transaction with the line row locked (allocations processed in
deterministic order-line order), so a 2-unit line can never refund 3
units or exceed its historical money. Credits-only orders get their own
auditable return path: allocations may exist without a payment_refund
(money forced to 0, actor required by CHECK) via
POST /orders/{id}/credit-refunds - no fake $0 monetary refunds.

+++ backend/app/schemas/payment.py
@@ -106,13 +106,13 @@ class TicketDelivery(ApiReadSchema):
 class TicketLineModifier(ApiReadSchema):
     group: str
     option: str
-    quantity: Decimal
+    quantity: int
     total: Decimal
 
 
 class TicketLine(ApiReadSchema):
     name: str
-    quantity: Decimal
+    quantity: int
     purchase_mode: str
     unit_price: Decimal
     line_total: Decimal
